% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapCCESS.R
\name{build_model}
\alias{build_model}
\title{build_model}
\usage{
build_model(num_features, num_hidden_features, num_latent_features)
}
\arguments{
\item{num_features}{number of input features}

\item{num_hidden_features}{number of hidden layer features}

\item{num_latent_features}{number of the features in embeddings}
}
\value{
a python object of deep learning model
}
\description{
Build the SnapCCESS VAE model
}
\examples{
#model=build_model(num_features=list(1000,49),
          #num_hidden_features=list(185,30),
          #num_latent_features=100)
}
