% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_SnapCCESS}
\alias{install_SnapCCESS}
\title{install_SnapCCESS}
\usage{
install_SnapCCESS(envname = "SnapCCESS", method = "conda", conda = "auto", ...)
}
\arguments{
\item{envname}{The name, or full path, of the environment in which Python packages are to be installed.
When NULL (the default), the active environment as set by the RETICULATE_PYTHON_ENV variable will be used;
if that is unset, then the r-reticulate environment will be used.}

\item{method}{Installation method. By default, "auto" automatically finds a method that
will work in the local environment. Change the default to force a specific installation method.
Note that the "virtualenv" method is not available on Windows.}

\item{conda}{The path to a conda executable.
By default, reticulate will check the PATH,
as well as other standard locations for Anaconda installations.}

\item{...}{Additional arguments passed to reticulate::py_install function}
}
\description{
install related python packages,
this should be run only once when you install the R pacakge.
}
\examples{
#install_SnapCCESS(envname = "SnapCCESS",method = "conda",conda="auto")

}
