% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadmodule.R
\name{loadmodule}
\alias{loadmodule}
\title{loadmodule}
\usage{
loadmodule()
}
\description{
Load all python modules that may used during analysis automatically,
including torch, snapccess, and numpy
}
\examples{
#loadmodule()

}
