% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapCCESS.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(x, mb_size = 64, num_workers = 0)
}
\arguments{
\item{x}{a list of input data modalities}

\item{mb_size}{mini batch size}

\item{num_workers}{Setting the argument num_workers as a positive integer will turn on
multi-process data loading with the specified number of loader worker processes.}
}
\value{
a python object loaded datasets
}
\description{
Data preprocess, PyTorch data loading utility
}
\examples{
#data=preprocess(lsit(rna,adt),mb_size=64)
}
