% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapCCESS.R
\name{run_SnapCCESS}
\alias{run_SnapCCESS}
\title{run_SnapCCESS}
\usage{
run_SnapCCESS(
  model,
  data,
  lr = 0.02,
  cycle = 50,
  epochs_per_cycle = 1,
  save_path = "",
  snapshot = TRUE,
  embedding_number = 1
)
}
\arguments{
\item{model}{SnapCCESS VAE model}

\item{data}{preprocessed dataset}

\item{lr}{initial learning rate}

\item{cycle}{number of epochs cycle, if snapshot is false,
this become to the total epochs that will be training in traditional VAE model}

\item{epochs_per_cycle}{number of epochs per cycle, if snapshot is false, this argument doesn't work}

\item{save_path}{which specifies the folder where the embedding will be saved.
If save_path is not specified or is an empty string, the embedding will only
be returned as an R object and will not be saved to the local folder.}

\item{snapshot}{Boolen value, True for run SnapCCESS, False will run traditional VAE}

\item{embedding_number}{the number in the filename to indicate that which embedding is saved to folder.}
}
\value{
a list of python objects, including model, loss of each epoch, and embeddings
}
\description{
Run the SnapCCESS or Traditional VAE trainning process to get the embeddings.
}
\examples{
#output=run_SnapCCESS(model,data,epochs=50,epochs_per_cycle=2,save_path="",snapshot=TRUE)
}
